findandbind: pass found entry to callback
Prior versions of `findandbind()` passed `(err, data)` to the callback.
This functionality was broken, I believe inadvertently, in commit
a4b9ce6e8eda61d1c86faf20995464d538b20f6c It's still documented in
the README and the callbacks in the unit tests still accept
`(err, data)`. Since it was broken in a minor version bump (3.0->3.1),
it will break existing code that relies on `data` being passed back
and uses the caret operator to specify dependency versions.

Restore that functionality, by passing the found entry
to the callback supplied to `findandbind()`.